requirements.txt
arabic-reshaper==2.1.4
asgiref==3.5.2
asn1crypto==1.5.1
automium==0.2.6
automium-web==0.1.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.1.2
django-crispy-forms==1.14.0
django-mathfilters==1.0.0
Flask==2.2.2
fonttools==4.37.4
fpdf2==2.5.7
future==0.18.2
html5lib==1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
legacy==0.1.6
lxml==4.9.1
MarkupSafe==2.1.1
numpy==1.23.4
oscrypto==1.3.0
pandas==1.5.0
Pillow==9.2.0
pycparser==2.21
pyHanko==0.15.0
pyhanko-certvalidator==0.19.5
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.11
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
svg.path==6.2
svglib==1.4.1
tabulate==0.9.0
tinycss2==1.1.1
tqdm==4.64.1
tzdata==2022.5
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.12
views==0.3
webencodings==0.5.1
Werkzeug==2.2.2
xhtml2pdf==0.2.8

